Refactor diff checks into rule registry
