+++ src/utils/collections.mjs
@@ -1,0 +1,7 @@
+export function uniqueSorted(values = []) {
+  return [...new Set(values)].sort();
+}
+
+export function formatList(values) {
+  return values && values.length > 0 ? values.join(", ") : "(none)";
+}
